backend/pyproject.toml
[project]
name = "synthdata-backend"
version = "0.1.0"
description = "SynthData Engine — Backend API"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=44.0.0",
    "python-dotenv>=1.2.0",
]

[project.scripts]
start = "uvicorn app.main:app --reload"

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.0.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["uv_build>=0.10.12,<0.11.0"]
build-backend = "uv_build"
